requirements.txt

PyQt5-sip==12.11.1
numpy==1.24.2
opencv-python==4.7.0.72
depthai==2.21.2.0

Pillow==9.4.0
requests==2.28.1
python-dotenv==0.21.0
